ffmpeg_sys_next

Function av_image_check_size

pub unsafe extern "C" fn av_image_check_size(
    w: c_uint,
    h: c_uint,
    log_offset: c_int,
    log_ctx: *mut c_void,
) -> c_int

Check if the given dimension of an image is valid, meaning that all bytes of the image can be addressed with a signed int.

@param w the width of the picture @param h the height of the picture @param log_offset the offset to sum to the log level for logging with log_ctx @param log_ctx the parent logging context, it may be NULL @return >= 0 if valid, a negative error code otherwise