ffmpeg_sys_next

Function av_small_strptime

pub unsafe extern "C" fn av_small_strptime(
    p: *const c_char,
    fmt: *const c_char,
    dt: *mut tm,
) -> *mut c_char

Simplified version of strptime

Parse the input string p according to the format string fmt and store its results in the structure dt. This implementation supports only a subset of the formats supported by the standard strptime().

The supported input field descriptors are listed below.

%%H: the hour as a decimal number, using a 24-hour clock, in the range ‘00’ through ‘23’

%%J: hours as a decimal number, in the range ‘0’ through INT_MAX

%%M: the minute as a decimal number, using a 24-hour clock, in the range ‘00’ through ‘59’

%%S: the second as a decimal number, using a 24-hour clock, in the range ‘00’ through ‘59’

%%Y: the year as a decimal number, using the Gregorian calendar

%%m: the month as a decimal number, in the range ‘1’ through ‘12’

%%d: the day of the month as a decimal number, in the range ‘1’ through ‘31’

%%T: alias for %%H:%%M:%%S

%%: a literal %

@return a pointer to the first character not processed in this function call. In case the input string contains more characters than required by the format string the return value points right after the last consumed input character. In case the whole input string is consumed the return value points to the null byte at the end of the string. On failure NULL is returned.