ffmpeg_sys_next

Function av_strtod

pub unsafe extern "C" fn av_strtod(
    numstr: *const c_char,
    tail: *mut *mut c_char,
) -> f64

Parse the string in numstr and return its value as a double. If the string is empty, contains only whitespaces, or does not contain an initial substring that has the expected syntax for a floating-point number, no conversion is performed. In this case, returns a value of zero and the value returned in tail is the value of numstr.

@param numstr a string representing a number, may contain one of the International System number postfixes, for example ‘K’, ‘M’, ‘G’. If ‘i’ is appended after the postfix, powers of 2 are used instead of powers of 10. The ‘B’ postfix multiplies the value by 8, and can be appended after another postfix or used alone. This allows using for example ‘KB’, ‘MiB’, ‘G’ and ‘B’ as postfix. @param tail if non-NULL puts here the pointer to the char next after the last parsed character